convert from cache to factory
